# docs/openapi-voice.yaml
openapi: 3.0.3
info:
  title: Xase Voice Data Governance API (MVP)
  version: 1.0.0
  description: API para governança de dados de voz. MVP com datasets, upload/webhook, policies, access, logs e ledger.
servers:
  - url: https://api.example.com
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
paths:
...
  /api/v1/datasets/{datasetId}:
    get:
      summary: Detalhar dataset
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: datasetId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
...
  /api/v1/datasets/{datasetId}/publish:
    post:
      summary: Publicar dataset (DRAFT -> ACTIVE)
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: datasetId
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
...